refactor(ui): fast-refresh compliance batch 5 — 7 route files, 3-6 warnings each
Part of #7850. 33 of 246 react-refresh/only-export-components warnings.

Same TanStack Router pattern as prior batches: moved each page's
component(s) to a new `-<route>-page.tsx` sibling, keeping only Route
(plus, where its config needs them, the search schema / validation) in
the route file.

admin-changes.index.tsx, blocks.index.tsx, extrinsics.index.tsx: the
now-familiar search-schema-stays / components-move split, each with a
second table component reading Route.useSearch()/fullPath swapped to
useSearch({from})/useNavigate({from}).

runtime.index.tsx: no search state, straight move of all 4 components.

extrinsics.$hash.tsx, validators.$hotkey.tsx: param-based routes (not
search-based) -- Route.useParams() swapped to useParams({from}).
validators.$hotkey.tsx's NominatorsSection additionally reads
Route.useSearch()/fullPath, same treatment.

status.tsx: RecentIncidents used Route.useSearch({select: (s) =>
s.window}) -- swapped to useSearch({from, select}), same selector
semantics.

Each notFoundComponent stayed inline in its route file (an anonymous
arrow function assigned directly to the config, not a named
declaration) -- the rule only flags identifiers, so these were never
part of the violation count and don't need moving.

Zero behavior change; no external importers of any moved component
existed outside its own route file.

230 -> 197 warnings; 0 eslint errors. Live-verified all 7 routes
(/admin-changes, /blocks, /extrinsics, /runtime, an extrinsic detail
page, /status, a validator detail page) -- render correctly with live
chain data, no console errors.
